"Renamed test_utils to utils to make room for tests of uq360.utils"
