+++ tests/test_ClassificationCalibration.py
@@ -4,7 +4,7 @@ from sklearn.metrics import brier_score_loss
 from sklearn.utils._testing import (
     assert_array_equal,
 )
-from tests.test_utils import create_train_test_prod_split, split
+from tests.utils import create_train_test_prod_split, split
 import unittest
 from unittest import TestCase
 from uq360.algorithms.classification_calibration import ClassificationCalibration
